Ensure data_mean and data_std are explicitly converted to float32 type in DedalusDataset2D class for improved consistency in data handling.

+++ utils/griddataset.py
@@ -1218,8 +1218,8 @@ class DedalusDataset2D(Dataset):
         # add timestep with 0 mean and 1 std 
         data_mean = np.concatenate([data_mean, np.zeros(1)], axis=-1)
         data_std = np.concatenate([data_std, np.ones(1)], axis=-1)
-        data_mean = torch.from_numpy(data_mean)
-        data_std = torch.from_numpy(data_std)
+        data_mean = torch.from_numpy(data_mean).astype(torch.float32)
+        data_std = torch.from_numpy(data_std).astype(torch.float32)
         return data_mean, data_std
     def downsample_x(self, u, N):
         """
